Before: `initial 20s incremental: 5.5s` 
Before with noEmitOnError=false manually patched-in: `initial 18s incremental: 3.15s`
After: `initial 7.7s incremental: 0.5s`

It is faster for the following two reasons: 
* no es6 
* no files from node_modules go into the browser tree.

Now, the time is split more-or-less equally between Funnel, MergeTrees, and TS. 

This PR also fixed the circular dependencies check that was broken.