@mary-poppins 2015-11-17
Merging PR #5299 on behalf of @vsavkin to branch [presubmit-vsavkin-pr-5299](https://github.com/angular/angular/tree/presubmit-vsavkin-pr-5299).